+++ lib/bindtree.ml
@@ -6,9 +6,9 @@ type binop =
 
 type expr =
   | Int of int64
-  | Var of string
+  | Var of string * int
   | BinOp of binop * expr * expr
 
 type t =
-  | Let of string * expr
+  | Let of string * int * expr
   | Expr of expr