+++ lib/report.ml
@@ -1,7 +1,13 @@
 type t = {
-  loc : Location.t;
+  loc : Location.t option;
   msg : string;
 }
 
-let make loc msg = { loc; msg }
-let show report = Location.report report.loc ^ ": " ^ report.msg
+let make_loc loc msg = { loc = Some loc; msg }
+and make msg = { loc = None; msg }
+
+and show report =
+  Option.fold ~none:""
+    ~some:(fun loc -> Location.report loc ^ ": ")
+    report.loc
+  ^ report.msg